java/src/test/java/com/yourtion/leetcode/easy/c9053/SolutionTest.java
package com.yourtion.leetcode.easy.c9053;

import com.yourtion.leetcode.TestUtils;
import com.yourtion.leetcode.utils.ListNode;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("面试题24. 反转链表")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[1,2,3,4,5]", "[5, 4, 3, 2, 1]")
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void reverseList(String source, String res) {
        System.out.printf("runTest: %s , res: %s", source, res);
        ListNode head = TestUtils.stringToListNode(source);
        ListNode ret = new Solution().reverseList(head);
        Assertions.assertEquals(res, TestUtils.listNodeToString(ret));
    }
}